backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.2
python-dotenv==1.0.0
openai==1.3.7
qdrant-client==1.6.0
psycopg2-binary==2.9.9
langchain==0.0.350 # For RecursiveCharacterTextSplitter and other utilities
